+++ OrbitGl/DataView.h
@@ -25,7 +25,7 @@ public:
     virtual const std::vector<std::wstring>& GetColumnHeaders();
     virtual const std::vector<float>& GetColumnHeadersRatios();
     virtual std::vector<std::wstring> GetContextMenu(int a_Index);
-    virtual int  GetNumElements() { return (int)m_Indices.size(); }
+    virtual size_t GetNumElements() { return m_Indices.size(); }
     virtual std::wstring GetValue(int /*a_Row*/, int /*a_Column*/) { return L""; }
     virtual std::wstring GetToolTip(int /*a_Row*/, int /*a_Column*/) { return L""; }
     virtual void SetFilter( const std::wstring & a_Filter ) { m_Filter = a_Filter; OnFilter( a_Filter ); }
